Factored out checkAllArgs, added tests
