+++ index.js
@@ -5,27 +5,19 @@ const stringLength = require('string-length')
 const channel = '#sprinthelper'
 const botName = 'sprint-helper1'
 
-const validateInput = require('./validate').validateInput
+const valid = require('./validate')
 
 const client = new irc.Client('irc.freenode.net', botName, {
     channels: [channel],
     port: process.env.PORT || 6667
 })
 
-function checkAllArgs (message) {
-  return (message.topic !== null &&
-    message.sprintIssue !== null &&
-    message.notes !== null &&
-    message.zoom !== null &&
-    message.stream !== null)
-}
-
 client.addListener('message', function (from, to, message) {
   message = parse(message)
 
   if (to === channel && message[0].slice(0, botName.length) === botName) {
-    message = validateInput(message)
-    if (checkAllArgs(message)) {
+    message = valid.validateInput(message)
+    if (valid.checkAllArgs(message)) {
       var header = `========================= IPFS Sprint: ${message.topic} =========================`
       client.say(channel, `
 ${header}
